Fix Overflow Vulnerability
This change fixes the vulnerability where creditManyAndOne could
credit the same account multiple times and skip accounts as a result
of uint256 to  uint16 overflow. We fix this by defaulting to uint256, the
max length of any EVM array at the expense of a small bump in gas.

+++ libs/coinflip-contracts/src/Wallets.sol
@@ -46,14 +46,14 @@ contract Wallets is UsingReentrancyGuard {
         address oneAccount,
         uint amountForOneAccount
     ) external payable {
-        uint16 manyAccountsLength = uint16(manyAccounts.length);
+        uint manyAccountsLength = manyAccounts.length;
         require(
             (amountForEachManyAccount * manyAccountsLength) +
                 amountForOneAccount ==
                 msg.value
         );
 
-        for (uint16 i; i < manyAccountsLength; ) {
+        for (uint i; i < manyAccountsLength; ) {
             _credit(manyAccounts[i], amountForEachManyAccount);
             unchecked {
                 ++i;
